changes to gcode
In order to send the instructions, you write to the gcode.txt file the function name that you wanna use from the algorithms.py file, and the positional arguments, evrything space seperated. To write the next function, you change line. In order to make this work, we should use the match keyword that is identical to the switch statement is other languages such as java, c++ etc. Then, to unpack the tuple of arguments we will use the * operator (*list_name). This operator will upack the list and turn the values into posistional arguments. All this functionalty will be inplemented on the test-gcode.py file.